moves schema config and normalizer importers to schema module, braks cyclic deps with dest capabilities
