+++ tests/normalize/test_max_nesting.py
@@ -8,7 +8,7 @@ from dlt.destinations import dummy
 from tests.common.utils import json_case_path
 
 
-TOP_LEVEL_TABLES = ["bot_events"]
+ROOT_TABLES = ["bot_events"]
 
 ALL_TABLES_FOR_RASA_EVENT = [
     "bot_events",
@@ -37,8 +37,8 @@ def rasa_event_bot_metadata():
 @pytest.mark.parametrize(
     "nesting_level,expected_num_tables,expected_table_names",
     (
-        (0, 1, TOP_LEVEL_TABLES),
-        (1, 1, TOP_LEVEL_TABLES),
+        (0, 1, ROOT_TABLES),
+        (1, 1, ROOT_TABLES),
         (2, 3, ALL_TABLES_FOR_RASA_EVENT_NESTING_LEVEL_2),
         (5, 8, ALL_TABLES_FOR_RASA_EVENT),
         (15, 8, ALL_TABLES_FOR_RASA_EVENT),
